pyproject.toml
[project]
name = "lab-log"
version = "1.0.0"
description = "实验室日志系统 MVP"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "PyMySQL>=1.1.0",
    "cryptography>=41.0.0",
    "dashscope>=1.17.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.1.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

